breaking data store into Engine (front end), Tables(individual table handling), and Store(table and overall data handling); moving index handling into Table
