+++ lib/greyscale_record/data_store/engine.rb
@@ -1,0 +1,70 @@
+module GreyscaleRecord
+  module DataStore
+    class Engine
+
+      # A data store can only have one driver. 
+      # It's like a database connection. If you want models to connect to different
+      #   databases, you have them inherit from different base classes that specify 
+      #   a db to connect to.
+
+      # Pros:
+      #   * All the data in your store comes from the same place. Easy to reason about
+      #   * No chance of data source collisions. Can't populate tables from multiple sources
+      #   * Cleaner table-adding interface 
+      #       store.add_table( name )
+      #            vs
+      #       store.add_table( name, driver )
+      #       # where does driver get initialized? 
+      #   * Straight forward patching: one complete patch at a time
+      #       How do you patch across differntly driven tables?
+      #       Do you need a patch stack?
+      #   * Depending on how your remote is built, you can pull in all the data at once
+      #
+      # Cons:
+      #   * Limiting if you want to have some tables come from YAML and some from remote
+      #
+      # Yeah. OK.
+
+      attr_reader :store
+      delegate :apply_patch, :remove_patch, :with_patch, to: :store 
+
+      def initialize( driver )
+        @driver = driver
+        @store  = Store.new
+      end
+
+      # Read only store. No writes allowed. 
+
+      def read( options = {} )
+        if GreyscaleRecord.live_reload
+          load_table!( name )
+        end
+
+        # TODO: this is where all the meat is
+        store[options[:_table]]
+      end
+
+      def add_table( name )
+        load_table! name
+      end
+
+      def add_index( table_name, column_name, options = {} )
+        store[table_name].add_index( column_name, options )
+      end
+
+      private
+
+      def store
+        @store
+      end
+
+      def load_table!( name )
+        store[name] = @driver.load!( name )
+      end
+
+      def patched?
+        !!Thread.current[:patched_data].present?
+      end
+    end
+  end
+end
